feat(runtime-fluurt): expose escapeAttrValue helper
