+++ packages/runtime/src/html/index.ts
@@ -1,4 +1,10 @@
-export { toString, escapeScript, escapeStyle, escapeXML } from "./content";
+export {
+  toString,
+  escapeScript,
+  escapeStyle,
+  escapeXML,
+  escapeAttrValue
+} from "./content";
 
 export { attr, attrs, classAttr, styleAttr } from "./attrs";
